refactor: parse
